Bryan: Filled in decent chunk of About Me
Added in text and some pictures. Also fixed previous nit things (namely
changing path names to be all lowercase)

L=ENG-122
