--- js/src/personal/bryan/AboutMe/AboutMe.module.css
@@ -1,45 +1,0 @@
-.title {
-  color: light-dark(var(--mantine-color-black), var(--mantine-color-white));
-  font-size: rem(64px);
-  letter-spacing: rem(-2px);
-
-  @media (max-width: $mantine-breakpoint-md) {
-    font-size: rem(50px);
-  }
-}
-
-.title2 {
-  color: light-dark(var(--mantine-color-black), var(--mantine-color-white));
-  font-size: rem(40px);
-  letter-spacing: rem(-2px);
-
-  @media (max-width: $mantine-breakpoint-md) {
-    font-size: rem(30px);
-  }
-}
-
-.position_school {
-  color: light-dark(var(--mantine-color-black), var(--mantine-color-white));
-  font-size: rem(48px);
-  letter-spacing: rem(-2px);
-
-  position: absolute;
-  top: 20rem;
-  left: 5rem;
-}
-
-.position_interest_title {
-  color: light-dark(var(--mantine-color-black), var(--mantine-color-white));
-  font-size: rem(48px);
-  letter-spacing: rem(-2px);
-
-  position: absolute;
-  top: 20rem;
-  left: 45rem;
-}
-
-.position_interest_text {
-  position: absolute;
-  top: 25rem;
-  left: 45rem;
-}
