--- js/src/personal/bryan/AboutMe/AboutMePictures.tsx
@@ -1,3 +1,0 @@
-export function AboutMePictures() {
-  return <>{'hello'}</>
-}
