+++ js/src/personal/bryan/Bryan.page.tsx
@@ -1,10 +1,10 @@
 import { MantineProvider, Flex } from '@mantine/core'
 import { Outlet } from 'react-router-dom'
 import { theme } from '@/personal/bryan/theme.ts'
-import { NavBar } from '@/personal/bryan/Frame/NavBar.tsx'
+import { NavBar } from '@/personal/bryan/frame/NavBar.tsx'
 import classes from '@/personal/bryan/Bryan.module.css'
-import { Socials } from '@/personal/bryan/Frame/Socials.tsx'
-import { TimeDisplay } from '@/personal/bryan/Frame/TimeDisplay.tsx'
+import { Socials } from '@/personal/bryan/frame/Socials.tsx'
+import { TimeDisplay } from '@/personal/bryan/frame/TimeDisplay.tsx'
 
 export function BryanPage() {
   return (
