+++ js/src/personal/bryan/aboutme/AboutMe.page.tsx
@@ -1,0 +1,93 @@
+import { Title, Group, Image, Space, Stack, Text, Anchor } from '@mantine/core'
+import classes from './AboutMe.module.css'
+import skiing from '@/personal/bryan/images/bryan_skiing.png'
+import mathey_logo from '@/personal/bryan/images/mathey_logo.jpeg'
+import tower_logo from '@/personal/bryan/images/tower_logo.jpeg'
+
+export function AboutMePage() {
+  return (
+    <>
+      <Title
+        className={classes.title}
+        component="span"
+        ml="5rem"
+      >
+        {'About Me'}
+      </Title>
+      <Group
+        justify="flex-start"
+        component="span"
+        ml="5rem"
+      >
+        <Stack>
+          <Image
+            h={310}
+            src={skiing}
+          />
+          <Title component="span">
+            {'cs @ princeton'}
+            <Space h="s" />
+            {'Class of 2026'}
+          </Title>
+        </Stack>
+        <Stack>
+          <Text component="span">{'Hi, I’m Bryan! I’m XYZ. '}</Text>
+          <Space />
+          <Text component="span">
+            {'I’m a head Copy editor for the Daily Princetonian'}
+          </Text>
+          <Anchor href="https://www.dailyprincetonian.com/">
+            {'The Daily Princetonian'}
+          </Anchor>
+          <Text component="span">
+            {
+              'I also lab TA for the Intro COS lab (specifically for COS 217/226)'
+            }
+          </Text>
+          <Text component="span">
+            {'Proud member of Mathey College and Tower Club'}
+          </Text>
+          <Group
+            justify="center"
+            mr="1.5rem"
+          >
+            <Image
+              h={100}
+              w={100}
+              src={mathey_logo}
+            />
+            <Space />
+            <Space />
+            <Space />
+            <Image
+              h={120}
+              w={120}
+              src={tower_logo}
+            />
+          </Group>
+        </Stack>
+        <Stack>
+          <Title
+            component="span"
+            className={classes.title2}
+          >
+            {'Interests'}
+          </Title>
+          <Text component="span">
+            {'I vibe with board games'}
+            <Space />
+            {'Also nature'}
+            <Space />
+            {'Also crosswords, but moreso making them'}
+          </Text>
+          <Anchor
+            target="_blank"
+            href="https://crossword.dailyprincetonian.com/"
+          >
+            {'Puzzles!'}
+          </Anchor>
+        </Stack>
+      </Group>
+    </>
+  )
+}
