+++ js/src/personal/bryan/projects/Projects.module.css
@@ -1,0 +1,19 @@
+.title {
+  color: light-dark(var(--mantine-color-black), var(--mantine-color-white));
+  font-size: rem(64px);
+  letter-spacing: rem(-2px);
+
+  @media (max-width: $mantine-breakpoint-md) {
+    font-size: rem(50px);
+  }
+}
+
+.title2 {
+  color: light-dark(var(--mantine-color-black), var(--mantine-color-white));
+  font-size: rem(40px);
+  letter-spacing: rem(-2px);
+
+  @media (max-width: $mantine-breakpoint-md) {
+    font-size: rem(30px);
+  }
+}
